requirements.txt
dash>=2.14.0
plotly>=5.17.0
pandas>=2.0.0
numpy>=1.24.0
fastapi>=0.100.0
uvicorn>=0.24.0
sqlalchemy>=2.0.0
requests>=2.31.0
python-dotenv>=1.0.0
dash-bootstrap-components>=1.5.0
